requirements.txt
psycopg2-binary ~= 2.9.4
pyqt5 ~= 5.15.7